707 Files , Zirnox finished , Arc welder , Solderingstation, RBMK Columms rendering (no logic) turbine texutre Issue fixed , Easteregg Item

+++ src/main/java/com/hbm_m/block/machines/UniversalMachinePartBlock.java
@@ -392,13 +392,15 @@ public class UniversalMachinePartBlock extends BaseEntityBlock implements IDeton
                     BlockState controllerState = pLevel.getBlockState(controllerPos);
                     // Проверяем, что контроллер ещё существует
                     if (controllerState.getBlock() instanceof IMultiblockController controller) {
-                        // Вызываем разрушение структуры
-                        if (controllerState.hasProperty(HorizontalDirectionalBlock.FACING)) {
-                            Direction facing = controllerState.getValue(HorizontalDirectionalBlock.FACING);
-                            controller.getStructureHelper().destroyStructure(pLevel, controllerPos, facing);
+                        if (!MultiblockStructureHelper.isDestroying() && controller.shouldDestroyOnPartRemoved()) {
+                            // Вызываем разрушение структуры
+                            if (controllerState.hasProperty(HorizontalDirectionalBlock.FACING)) {
+                                Direction facing = controllerState.getValue(HorizontalDirectionalBlock.FACING);
+                                controller.getStructureHelper().destroyStructure(pLevel, controllerPos, facing);
+                            }
+                            // Ломаем контроллер
+                            pLevel.destroyBlock(controllerPos, true);
                         }
-                        // Ломаем контроллер
-                        pLevel.destroyBlock(controllerPos, true);
                     }
                 }
             }
